Moving ecs definitions to a cpp file
